% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuRnip.R
\name{load_raw}
\alias{load_raw}
\title{Load raw data}
\usage{
load_raw(filename, wdir = here::here(), col_names = FALSE)
}
\arguments{
\item{filename}{String with the input filename.}

\item{wdir}{String to working directory.}

\item{col_names}{Boolean flag to indicate whether the input file contains
column names.}
}
\value{
Tibble structure with raw data.
}
\description{
Load raw data from CSV file. This file should have two columns:
\enumerate{
\item \code{datetime}: Formatted as \code{yyyy-MM-dd HH:mm}.
\item \code{price}: Integer.
}
}
